test: 상품 정보 변경 테스트

+++ preorderService/itemService/src/main/java/com/preOrderService/dto/ItemResponseDto.java
@@ -1,7 +1,6 @@
 package com.preOrderService.dto;
 
 import com.preOrderService.entity.ItemType;
-import jakarta.persistence.Enumerated;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -12,12 +11,12 @@ public class ItemResponseDto {
     private Long id;
     private String name;
     private String content;
-    private int price;
-    private int stock;
+    private Long price;
+    private Long stock;
     private LocalDateTime reserveTime;
     private ItemType type;
     @Builder
-    public ItemResponseDto(Long id, String name, String content, int price, int stock, LocalDateTime reserveTime, ItemType type) {
+    public ItemResponseDto(Long id, String name, String content, Long price, Long stock, LocalDateTime reserveTime, ItemType type) {
         this.id = id;
         this.name = name;
         this.content = content;
